requirements.txt
numpy
pandas
seaborn
matplotlib
scikit-learn
python-dotenv
flask
pymongo
#-e .